:sparkles: Exceptions differently
