+++ config.py
@@ -5,4 +5,4 @@ FPS = 30
 CHANNELS = 32
 UPDATE_RATE_MS = 50
 INIT_CHANNEL, INIT_CHANNEL_VALUE = 1, 100
-LOGLVL = logging.INFO
+LOGLVL = logging.DEBUG
