+++ templates/index.html
@@ -2,7 +2,7 @@
 {% block title %}This is an example page{% endblock %}
 
 {% block content %}
-  <h1>Hello</h1>
+  <h1>Animation</h1>
     <div class="container-fluid">
         <div class="row">
             <div class="col-md-4">
@@ -12,12 +12,12 @@
                         <input type="color" name="color" class="form-control" id="color" placeholder="#8f3596">
                     </div>
                     <div class="form-group">
-                        <label for="duration">Duration</label>
-                        <input type="number" name="duration" class="form-control" id="duration" placeholder=10000>
+                        <label for="duration">Duration (ms)</label>
+                        <input type="number" name="duration" class="form-control" id="duration" value="2000" max="100000">
                     </div>
                     <div class="form-group">
                         <label for="ease">Ease (see <a href="https://github.com/asweigart/pytweening">PyTweening</a>)</label>
-                        <input type="text" name="ease" class="form-control" id="ease" placeholder="linear">
+                        <input type="text" name="ease" class="form-control" id="ease" value="linear">
                     </div>
                     <button type="submit" class="btn btn-primary">Submit</button>
                 </form>
